requirements.txt
psutil
gpuinfo
xformers
accelerate
transformers
diffusers
fairscale
ftfy
huggingface-hub
Pillow
timm
open_clip_torch
requests
clip-interrogator
torch
runpod
spacy
opencv-python
flask
flask_cors
pandas
gfpgan
realesrgan
gunicorn